09/05/2022 Looping and branching added.
Added equipment loop.
Added game loop.
All branching and looping have error checking.
The inputs check upper and lower case letters in the first switch.  After that, the input chars are changed to uppercase.  That is done in two different ways.  Once the input is changed to uppercase in a separate command.  The other times, the char is converted to uppercase on the same line of code that gets the input and converts it from a String to a char.