09/05/2022 Started making the RPG object oriented.
Added NPC class (with enum Races).
Added Player class (with enum Equipment).
Realized that I do not need to have EquipmentName because the enum can be printed directly.  If you want to print the number, you need to typecast it as an int.  (That is a cool thing about enum in C#.)