09/05/2022 Added NPC interaction.
The NPC interaction is not the greatest, but it is interaction.  It could be better and based on the damage points of the Player and NPC, but I wanted to get something that worked without much fuss.  And I wanted to give just a sample, without trying to dictate how things ought to be.